+++ app/Console/Commands/Build.php
@@ -59,7 +59,7 @@ class Build extends Command
     /**
      * Execute the console command.
      */
-    public function handle(): void
+    public function fire(): void
     {
         if (Phar::canWrite()) {
             $this->build($this->input->getArgument('name') ?: self::BUILD_NAME);